# src/main/resources/api.yaml
openapi: 3.0.3

info:
  title: Loan Request API
  description: API for managing personal loan requests
  version: 1.0.0

servers:
  - url: http://localhost:8080

tags:
  - name: Loans
    description: Loan request operations

paths:
  /loans:
    post:
      tags: [Loans]
      summary: Create new loan request
      operationId: createLoan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanDTO'
      parameters:
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Loan created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanDTO'
        '400':
          description: Invalid input data

    get:
      tags: [Loans]
      summary: Get all loan requests
      operationId: getLoans
      responses:
        '200':
          description: List of loan requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoanDTO'

  /loans/{id}:
    get:
      tags: [Loans]
      summary: Get loan request by ID
      operationId: getLoanById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Loan found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanDTO'
        '404':
          description: Loan not found

  /loans/{id}/status:
    patch:
      tags: [Loans]
      summary: Update loan status
      operationId: updateLoanStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  description: New status of the loan
                  enum: [ PENDING, APPROVED, REJECTED, CANCELLED ]
                  example: APPROVED
      responses:
        '200':
          description: Loan status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanDTO'
        '400':
          description: Invalid status transition
        '404':
          description: Loan not found

  /auth/me:
    get:
      summary: Show the user information
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'

  /users:
    post:
      tags: [Users]
      summary: Create new user
      description: Only managers can create users
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '403':
          description: Forbidden

components:
  schemas:
    LoanDTO:
      type: object
      required:
        - applicantName
        - amount
        - currency
        - documentId
      properties:
        id:
          type: string
          format: uuid
        applicantName:
          type: string
          example: Jane Doe
        amount:
          type: number
          format: double
          example: 10000.00
        currency:
          type: string
          example: EUR
        documentId:
          type: string
          example: 12345678A
        createdAt:
          type: string
          format: timestamp
        status:
          $ref: '#/components/schemas/LoanStatus'
        createdBy:
          type: string
          format: uuid

    UserDTO:
      type: object
      required:
        - username
        - role
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          example: John Doe
        role:
          type: string
          description: Role of the user
          enum: [ CLIENT, MANAGER ]
          example: CLIENT


    LoanStatus:
      type: string
      enum:
        - PENDING
        - APPROVED
        - REJECTED
        - CANCELLED